+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This projec
 - **unreleased**
   - Improve performance of `scene.select()`.
   - Improve performance of `scene.set()`, in the case where the component type is already present and the component will be replaced.
+  - Improve performance of `scene.collect()`.
   - Improve performance of internal methods (improving overall performance).
 
   - Other changes: